feat: moving explore dash to a separate feature module
