+++ common/src/main/res/values/attrs.xml
@@ -17,4 +17,9 @@
     <declare-styleable name="FingerprintView">
         <attr name="custom_layout" format="reference" />
     </declare-styleable>
+
+    <declare-styleable name="VectorColors">
+        <attr name="primaryColor" format="color" />
+        <attr name="secondaryColor" format="color" />
+    </declare-styleable>
 </resources>
