+++ features/exploredash/src/main/java/org/dash/wallet/features/exploredash/ExploreFragment.kt
@@ -1,11 +1,10 @@
-package de.schildbach.wallet.ui.explore
+package org.dash.wallet.features.exploredash
 
 import android.os.Bundle
 import android.view.View
 import androidx.fragment.app.Fragment
-import de.schildbach.wallet_test.R
-import de.schildbach.wallet_test.databinding.FragmentExploreBinding
 import org.dash.wallet.common.ui.viewBinding
+import org.dash.wallet.features.exploredash.databinding.FragmentExploreBinding
 
 class ExploreFragment : Fragment(R.layout.fragment_explore) {
