+++ settings.gradle
@@ -4,3 +4,4 @@ include 'sample-integration-android'
 include ':uphold-integration'
 include ':common'
 include ':liquid-integration'
+include ':features:exploredash'
