+++ wallet/build.gradle
@@ -14,13 +14,18 @@ repositories {
 }
 
 dependencies {
-    implementation 'com.google.android.material:material:1.3.0'
-    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
+    implementation project(path: ':common')
+    implementation project(path: ":features:exploredash")
+    //Liquid Integration
+    implementation project(path: ':liquid-integration')
+    //Uphold Integration
+    implementation project(path: ':uphold-integration')
     implementation(project(':integration-android')) {
         exclude(group: 'com.google.android', module: 'android')
     }
-    //Liquid Integration
-    implementation project(path: ':liquid-integration')
+
+    implementation 'com.google.android.material:material:1.3.0'
+    implementation 'androidx.legacy:legacy-support-v4:1.0.0'
 
     implementation 'androidx.annotation:annotation:1.2.0'
     implementation 'androidx.core:core-ktx:1.5.0'
@@ -55,17 +60,13 @@ dependencies {
     implementation 'com.github.tony19:logback-android-core:1.1.1-6'
     implementation 'com.google.code.findbugs:jsr305:3.0.2'
     implementation 'com.tbuonomo.andrui:viewpagerdotsindicator:4.1.2'
-
-    implementation project(path: ':common')
     implementation 'com.google.android.gms:play-services-auth:19.0.0'
+    implementation 'androidx.navigation:navigation-fragment-ktx:2.3.5'
 
     def room_version = "2.3.0"
     implementation "androidx.room:room-ktx:$room_version"
     implementation "androidx.room:room-runtime:$room_version"
     annotationProcessor "androidx.room:room-compiler:$room_version"
-
-    //Uphold Integration
-    implementation project(path: ':uphold-integration')
 
     implementation 'androidx.multidex:multidex:2.0.1'
