+++ wallet/res/layout/activity_explore.xml
@@ -1,7 +1,17 @@
 <?xml version="1.0" encoding="utf-8"?>
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:id="@+id/container"
-    tools:context="de.schildbach.wallet.ui.explore.ExploreActivity"/>
+    tools:context="de.schildbach.wallet.ui.explore.ExploreActivity">
+
+    <fragment
+        android:id="@+id/main_fragment_host"
+        android:name="androidx.navigation.fragment.NavHostFragment"
+        android:layout_width="match_parent"
+        android:layout_height="match_parent"
+        app:defaultNavHost="true"
+        app:navGraph="@navigation/explore_dash" />
+</FrameLayout>
