+++ wallet/res/values/attrs.xml
@@ -42,9 +42,4 @@
         <attr name="src" format="reference" />
         <attr name="text" format="reference" />
     </declare-styleable>
-
-    <declare-styleable name="VectorColors">
-        <attr name="primaryColor" format="color" />
-        <attr name="secondaryColor" format="color" />
-    </declare-styleable>
 </resources>
